Add pinned posts, fix hr styling, made bool envs
I added a pinned posts modelfield to articles and send all
pinned articles to the pinned posts section of the home view.
I added some logic to create cards from the pinned posts. I also
fixed an hr styling issue with the homeview to correct spacing. I
changed some of the methods that run in the post save signal to
be ignored if there is no city or country. I also changed the S3
and ADMIN_ENABLED envs to be booleans all the time. Seemed to fix
my issues.