Update home, about. Showing X posts. New JS
New home and about messages to include other kinds of posts.
I added the JS to count the number of posts being shown to the
allArticles view. The JS is contained in a static file of the
same name. I also added buttons to return to all posts from an
article. I added a search results header tag in the search view.
I added a all posts header tag in the all articles view.