Fix full-text search duplicate return issue
So I had to concatenate the tags into a single string that
then is passed to the vector which normalizes it. What I dont
understand is if the SearchVector finds lexemes for a given word
and it handles duplicates, why doesn't it handle mine? I'm thinking
I need to make writeups of interesting problems I encounter along the
way.