Add contact app to django, remove indent from site
I added the contact app to django and the model. The urls, sitemaps,
views, and admin python files are done. The app works with the new
contact html page that is linked in the navbar and footer. In the
future I will add reCAPTCHA and some filters to protect my contact
table in the database from spam entries.