Add email case fix to contact, subscriber forms
Now it wont add multiples of the same email. Also I lowered the
reCAPTCHA threshold to 0.5. I tired to add it to the models but
gave up. Decided to stick with the visible badge on pages with
forms as well. Changed the name of the field that will hold a
history of any exchanges between me an contacted parties to
"responses" from "responded".