Remove names from subscriber model
I dont need the names so I removed them from the model, the html,
and the views. Maybe I missed them in the styling. I fixed the
letter tehmplate having the wrong link aswell. It sends people to
the unsubscribe link as well now. I also changed the text drop down
in the letter model for FK to show the article slug.