Recaptcha sillyness read below
Basically I had a distinct version of django-recaptcha installed
on my machines. On the desktop PC the packaged is referenced as
captcha in the settings.py and on my ubuntu laptop its called
django-recaptcha. I settled it all by installing version 4.0.0 of
this packaged and everything is now on the same page. It looks like
pip install django-recaptcha defers to an older build tho so I have
to make sure Im specific on my computers.